+++ packages/router/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @virentia/router
 
+## 0.3.1
+
+### Patch Changes
+
+- chore: bump virentia packages
+
 ## 0.3.0
 
 ### Minor Changes